requirements.txt
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
wandb>=0.16.0
albumentations>=1.3.0
pytorch-fid>=0.3.0
scikit-learn>=1.3.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
tqdm>=4.66.0
rich>=13.5.0
tensorboard>=2.14.0
timm>=0.9.0
omegaconf>=2.3.0
pyyaml>=6.0
pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0
opencv-python>=4.8.0
requests>=2.31.0
lpips>=0.1.4
